Update README.md
The code is written in C programming language and consists of several header files and functions. The program includes a menu with several options, including starting the game, viewing the highest score, resetting the score, and exiting the game. The program also prompts the user to enter their name before starting the quiz game. It stores the highest score in a separate file and displays it when the user selects the option to view the highest score. The code is designed to be user-friendly and easy to navigate